Split providers and grant config to a new files
